Обновил на стриме :+1:

+++ src/main/java/ru/readysetcock/fate_telegram_bot/repository/TaroCardOfTheDayRepository.java
@@ -5,7 +5,7 @@ import org.springframework.stereotype.Repository;
 import ru.readysetcock.fate_telegram_bot.model.domain.TaroCardOfTheDay;
 
 @Repository
-public interface TaroCardOfTheDayRepository extends CrudRepository<TaroCardOfTheDay, Integer> {
+public interface TaroCardOfTheDayRepository extends CrudRepository<TaroCardOfTheDay, Long> {
     Boolean existsByUserId(Long id);
     TaroCardOfTheDay findTaroCardOfTheDayByUserId(Long id);
 }
